+++ src/load.c
@@ -166,7 +166,7 @@ read_irep_record_1(mrb_state *mrb, const uint8_t *bin, size_t *len, mrb_bool all
 
   diff = src - bin;
   mrb_assert(diff >= 0);
-  mrb_assert((size_t)diff <= SIZE_MAX);
+  mrb_assert(diff <= SIZE_MAX);
   *len = (size_t)diff;
 
   return irep;
@@ -350,7 +350,7 @@ read_debug_record(mrb_state *mrb, const uint8_t *start, mrb_irep* irep, size_t *
 
   diff = bin - start;
   mrb_assert(diff >= 0);
-  mrb_assert((size_t)diff <= SIZE_MAX);
+  mrb_assert(diff <= SIZE_MAX);
 
   if (record_size != (size_t)diff) {
     return MRB_DUMP_GENERAL_FAILURE;
@@ -367,7 +367,7 @@ read_debug_record(mrb_state *mrb, const uint8_t *start, mrb_irep* irep, size_t *
 
   diff = bin - start;
   mrb_assert(diff >= 0);
-  mrb_assert((size_t)diff <= SIZE_MAX);
+  mrb_assert(diff <= SIZE_MAX);
   *record_len = (size_t)diff;
 
   return MRB_DUMP_OK;
